test(executor): extract shared executor test infra from defer tests
- Create some helpers that we're going to reuse
- Pure relocation plus rename; no behavior change to the defer tests.
